Bugfix: queue randomly stopping.
Theory for problem - thread was not terminating / exiting at the end of trial procedure. Therefore extra thread being generated on each trial. After some number of trials there are too many threads to run simultaneously (?). The number of trials that the queue crashed after was random due to other background threads running in the OS being created and destroyed.

Fixed initially by calling thread.terminate() on finish_trial and on stop_queue such that thread is always terminated before next trial is started and the thread is restarted. This led to crashing and consulting the docs revealed that this function is unstable as it terminates the thread at any point in execution. Apparently thread.exit() is safer so used this instead. So far no queue crashes with this approach but have not tested with pause queue.